Deal with missing fields better
Previously, crajure assumed that all listings had all fields (such as
region). Unfortunately, they're actually optional. This patch alters the
behavior of the listing parser so that logical listings are processed
one at a time rather than running a CSS selector, praying that the
cardinalities match up and blindly zipping.

The result is that the region selector and others are much more useful
because they're guaranteed to match up to the appropriate listing rather
than being random.